+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/atom_call_follows_usage.ex
@@ -1,0 +1,8 @@
+defmodule Prefix.AtomCallFollowsUsage do
+  alias Prefix.PublicFunctionDeclaration
+
+  def run do
+    PublicFunctionDeclaration.<caret>
+    :maps.values(%{})
+  end
+end