+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/defimpl_name_usage.ex
@@ -1,0 +1,2 @@
+defimpl Prefix.PublicFunctionDeclaration.<caret> do
+end