+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/defmodule_name_do_block_usage.ex
@@ -1,0 +1,2 @@
+defmodule Prefix.PublicFunctionDeclaration.<caret> do
+end