+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/defmodule_name_usage.ex
@@ -1,0 +1,1 @@
+defmodule Prefix.PublicFunctionDeclaration.<caret>