+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/import_directive_usage.ex
@@ -1,0 +1,3 @@
+defmodule Prefix.ImportDirectiveUsage do
+  import Prefix.PublicFunctionDeclaration.<caret>
+end