+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/nested_qualifier_usage.ex
@@ -1,0 +1,6 @@
+defmodule Prefix.NestedQualifierUsage do
+  def run do
+    Prefix.PublicFunctionDeclaration.<caret>
+    Prefix.PublicFunctionDeclaration.public_function1()
+  end
+end