+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/require_directive_usage.ex
@@ -1,0 +1,3 @@
+defmodule Prefix.RequireDirectiveUsage do
+  require Prefix.PublicFunctionDeclaration.<caret>
+end