+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/same_file_enclosing_usage.ex
@@ -1,0 +1,11 @@
+defmodule Outer do
+  def before_function(a), do: a
+
+  def run do
+    Outer.<caret>
+  end
+
+  def after_function(a), do: a
+
+  defdelegate delegated(a), to: Enum
+end