+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/same_file_nested_after_caret_usage.ex
@@ -1,0 +1,9 @@
+defmodule Outer do
+  def run do
+    Inner.<caret>
+  end
+
+  defmodule Inner do
+    def inner_function(a), do: a
+  end
+end