+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/same_file_nested_usage.ex
@@ -1,0 +1,10 @@
+defmodule Outer do
+  defmodule Inner do
+    def inner_function(a), do: a
+  end
+
+  def run do
+    Inner.<caret>
+    Inner.inner_function(1)
+  end
+end