+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/struct_literal_no_braces_usage.ex
@@ -1,0 +1,5 @@
+defmodule Prefix.StructLiteralNoBracesUsage do
+  def run do
+    %Prefix.PublicFunctionDeclaration.<caret>
+  end
+end