+++ testData/org/elixir_lang/code_insight/completion/contributor/call_definition_clause/use_no_arguments_do_block_usage.ex
@@ -1,0 +1,7 @@
+defmodule Prefix.UseNoArgumentsDoBlockUsage do
+  alias Prefix.PublicFunctionDeclaration
+
+  use do
+    PublicFunctionDeclaration.<caret>
+  end
+end